+++ shv/rpctypes/blob.py
@@ -62,10 +62,12 @@ class RpcTypeBlob(RpcType):
                 lim = f"({self._min or ''},{self._max or ''})"
         return f"x{lim}"
 
-    def is_valid(self, value: SHVType) -> typing.TypeGuard[bytes]:  # noqa: D102
-        return self.validate(value) is None
+    def is_valid(  # noqa: D102
+        self, value: SHVType, is_updatable: bool = False
+    ) -> typing.TypeGuard[bytes]:
+        return self.validate(value, is_updatable) is None
 
-    def validate(self, value: SHVType) -> str | None:  # noqa: D102
+    def validate(self, value: SHVType, is_updatable: bool = False) -> str | None:  # noqa: D102
         if not isinstance(value, bytes):
             return "expected Blob"
         if self._min is not None and len(value) < self._min: