+++ shv/rpctypes/enum.py
@@ -42,10 +42,12 @@ class RpcTypeEnum(RpcType, collections.abc.Mapping[int, str]):
     def __len__(self) -> int:
         return len(self._items)
 
-    def is_valid(self, value: SHVType) -> typing.TypeGuard[int]:  # noqa: D102
-        return super().is_valid(value)
+    def is_valid(  # noqa: D102
+        self, value: SHVType, is_updatable: bool = False
+    ) -> typing.TypeGuard[int]:
+        return super().is_valid(value, is_updatable)
 
-    def validate(self, value: SHVType) -> str | None:  # noqa: D102
+    def validate(self, value: SHVType, is_updatable: bool = False) -> str | None:  # noqa: D102
         if not isinstance(value, int):
             return "expected Integer(Enum)"
         if value not in self: