+++ shv/rpctypes/keystruct.py
@@ -32,18 +32,21 @@ class RpcTypeKeyStruct(RpcType, collections.abc.Mapping[str, RpcType]):
     def __len__(self) -> int:
         return len(self._items)
 
-    def is_valid(self, value: SHVType) -> typing.TypeGuard[SHVMapType]:  # noqa: D102
-        return self.validate(value) is None
+    def is_valid(  # noqa: D102
+        self, value: SHVType, is_updatable: bool = False
+    ) -> typing.TypeGuard[SHVMapType]:
+        return self.validate(value, is_updatable) is None
 
-    def validate(self, value: SHVType) -> str | None:  # noqa: D102
+    def validate(self, value: SHVType, is_updatable: bool = False) -> str | None:  # noqa: D102
         if not is_shvmap(value):
             return "expected KeyStruct"
         if invalid := set(value) - set(self._items):
             return f"undefined KeyStruct key: {', '.join(str(v) for v in invalid)}"
         for k, item in self.items():
             val = value.get(k, None)
-            if (msg := item.validate(val)) is not None:
-                return f"invalid KeyStruct item {k}: {msg}"
+            if val is not None or not is_updatable:  # Allow None on update
+                if (msg := item.validate(val, is_updatable)) is not None:
+                    return f"invalid KeyStruct item {k}: {msg}"
         return None
 
     def inflate(self, value: SHVType) -> SHVMapType:  # noqa: D102