+++ shv/rpctypes/map.py
@@ -35,14 +35,16 @@ class RpcTypeMap(RpcType):
     def __str__(self) -> str:
         return f"{{{self._tp}}}"
 
-    def is_valid(self, value: SHVType) -> typing.TypeGuard[SHVMapType]:  # noqa: D102
-        return self.validate(value) is None
+    def is_valid(  # noqa: D102
+        self, value: SHVType, is_updatable: bool = False
+    ) -> typing.TypeGuard[SHVMapType]:
+        return self.validate(value, is_updatable) is None
 
-    def validate(self, value: SHVType) -> str | None:  # noqa: D102
+    def validate(self, value: SHVType, is_updatable: bool = False) -> str | None:  # noqa: D102
         if not is_shvmap(value):
             return "expected Map"
         for k, val in value.items():
-            if (msg := self._tp.validate(val)) is not None:
+            if (msg := self._tp.validate(val, is_updatable)) is not None:
                 return f"invalid Map item {k}: {msg}"
         return None