+++ shv/rpctypes/oneof.py
@@ -50,10 +50,10 @@ class RpcTypeOneOf(RpcType):
     def __str__(self) -> str:
         return "|".join(str(t) for t in self._types)
 
-    def validate(self, value: SHVType) -> str | None:  # noqa: D102
+    def validate(self, value: SHVType, is_updatable: bool = False) -> str | None:  # noqa: D102
         msgs = []
         for tp in self._types:
-            if (msg := tp.validate(value)) is not None:
+            if (msg := tp.validate(value, is_updatable)) is not None:
                 msgs.append(msg)
             else:
                 return None