+++ shv/rpctypes/standard.py
@@ -46,8 +46,8 @@ class RpcTypeStandard(RpcType):
             and self._tp == other._tp
         )
 
-    def validate(self, value: SHVType) -> str | None:  # noqa: D102
-        return self._tp.validate(value)
+    def validate(self, value: SHVType, is_updatable: bool = False) -> str | None:  # noqa: D102
+        return self._tp.validate(value, is_updatable)
 
 
 rpctype_dir = RpcTypeStandard(